requirements.txt
# --- Core ---
fastapi
uvicorn[standard]
python-socketio
python-multipart

# --- Database & ODM ---
motor
beanie          # <--- Mới thêm cái này
dnspython       # Bắt buộc để kết nối MongoDB Atlas

# --- Utilities ---
python-dotenv
pydantic
pydantic-settings
requests
cloudinary==1.44.1
aiofiles
jinja2